% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\name{as.LOBdataset}
\alias{as.LOBdataset}
\title{Transformation of a single longitudinal data matrix into LoBrA Data Object.}
\usage{
as.LOBdataset(longData, name = "", id = "id", time = "time",
  type = "type", class = "class", bg = FALSE)
}
\arguments{
\item{longData}{Matrix of longitudinal data containing all components}

\item{name}{name of the dataset}

\item{id}{name to identify the experiment id column}

\item{time}{name to identify the time column}

\item{type}{name to identify the type column}

\item{class}{name to identify the class column}

\item{bg}{indicates whether the data table contains background data}
}
\value{
LoBra data object
}
\description{
Real signals and background noise originating from experimental settings or random events
}
\examples{
\dontrun{
} 
  data(longDataExample)
  id="id"
  name="Longitudinal Test Dataset"
  time="time"
  type="type"
  class="class"
  bg=FALSE
  longData<-longDataExample
  ldo<-as.LOBdataset(longDataExample, name, bg=TRUE)
  save(ldo,file="data/ldoExample.RData")
  
}
