% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{createBGComponents}
\alias{createBGComponents}
\title{Simulate background noise peaks}
\usage{
createBGComponents(
  components,
  samples,
  labels,
  timepoints = 15,
  bg = 3,
  mean = 5,
  sd = 3,
  experimentSD = 2,
  randomnoise = 0.1,
  plotting = FALSE
)
}
\arguments{
\item{components}{number of background components to be created}

\item{samples}{number of experiments}

\item{labels}{name of each experiment}

\item{timepoints}{number of sample measurements}

\item{bg}{number of background measurements}

\item{mean}{mean value of noise components}

\item{sd}{standard deviation value of noise for this component}

\item{experimentSD}{standard deviation value of each experiment for this component}

\item{randomnoise}{random variation changing at each time point}

\item{plotting}{Indicator whether the component will be plotted (TRUE) or not (FALSE)}
}
\value{
matrix of background components
}
\description{
Simulating background noise signals originating from experimental settings or random events
}
