% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{createBGData}
\alias{createBGData}
\title{Create example background measurements}
\usage{
createBGData(samples = 10, bg = 3, mean = 0, sd = 1,
  randomnoise = 0.1)
}
\arguments{
\item{samples}{number of experiments}

\item{bg}{number of background measurements}

\item{mean}{mean value of noise for this component}

\item{sd}{standard deviation value of noise for this component}

\item{randomnoise}{random variation changing at each time point}
}
\value{
matrix of background measurements
}
\description{
Background noise signals originating from experimental settings or random events
}
