% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{GaudermanLDO-class}
\alias{GaudermanLDO-class}
\alias{GaudermanLDO}
\title{An S4 class to represent a 'Gouderman' LDO object, that was generated by the generalized gauderman algorithm.}
\description{
An S4 class to represent a 'Gouderman' LDO object, that was generated by the generalized gauderman algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{character Name of the new 'generalized-Gauderman' adjusted longitudinal data}

\item{\code{dataFrames}}{list List of 'generalized-Gauderman' modified data. One \code{data.frame} for each component.}

\item{\code{peaknames}}{character Vector of component names contained in this object.}

\item{\code{k}}{numeric Updated times for the breaks of the spline model.}

\item{\code{times}}{matrix Vector of updated time values.}

\item{\code{newTimeVars}}{character The names of the newly defined time variables of the generalized 'Gauderman' model.}

\item{\code{ids}}{character Vector of identifiers for the experiments}

\item{\code{labels}}{factor Vector of class labels for each experiment}
}}

