% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{GaudermanModelEvaluation-class}
\alias{GaudermanModelEvaluation-class}
\alias{GaudermanModelEvaluation}
\title{An S4 class to represent the result of the linear mixed effect modeling on a gauderman LDO.}
\description{
An S4 class to represent the result of the linear mixed effect modeling on a gauderman LDO.
}
\section{Slots}{

\describe{
\item{\code{name}}{character Name of the new 'generalized-Gauderman' adjusted longitudinal model.}

\item{\code{gaudermanLDO}}{GaudermanLDO 'Generalized-Gauderman' adjusted longitudinal data object.}

\item{\code{models}}{list List of models generated for each component.}

\item{\code{labels}}{factor Vector of class labels for each experiment}

\item{\code{pvalues}}{matrix Matrix of p-values for the intercept as well as all slops of the spline model for each component.}

\item{\code{correctedpvalues}}{matrix Matrix of corrected p-values for the intercept as well as all slops of the spline model for each component.}

\item{\code{modelparameter}}{matrix Model parameter for each component.}
}}

