% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{getColor}
\alias{getColor}
\title{Get colors for the plotting function.}
\usage{
getColor(label, size)
}
\arguments{
\item{label}{class labels of the samples}

\item{size}{size of the color vector to be created}
}
\value{
col vector of colors created
}
\description{
Get colors for the plotting function.
}
