% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{getPvalue}
\alias{getPvalue}
\title{Testing differences of groups with respect to a specific value and test.}
\usage{
getPvalue(y, group, test)
}
\arguments{
\item{y}{Values to be tested}

\item{group}{corresponding groups whose difference we want to test}

\item{test}{specific test to be used. Can be each of the following 'bf', 'levene' or 'bartlett'.}
}
\description{
Testing differences of groups with respect to a specific value and test.
}
