% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{LDO-class}
\alias{LDO-class}
\alias{LDO}
\title{An S4 class to represent a 'LoBrA' Data Object (LDO). It stores multiple time series data for muliple experiements and multiple Components. It allows repeated measurements of a component, irregular sampling, and unequal temporal spacing of the time points.}
\description{
An S4 class to represent a 'LoBrA' Data Object (LDO). It stores multiple time series data for muliple experiements and multiple Components. It allows repeated measurements of a component, irregular sampling, and unequal temporal spacing of the time points.
}
\section{Slots}{

\describe{
\item{\code{name}}{character Name of the 'LDO' object}

\item{\code{dataMatrices}}{list List of matrices of component measurement. It contains a measurement for each time point and each experiment.}

\item{\code{backgroundMatrices}}{list List of matrices of background measurements. It contains a measurement for each time point and each experiment.}

\item{\code{peaknames}}{character Character vector of Component names}

\item{\code{times}}{numeric Vector of times for each time point in the data}

\item{\code{ids}}{character Vector of identifiers for the experiments}

\item{\code{labels}}{factor Vector of class labels for each experiment}
}}

