% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ldo}
\alias{ldo}
\title{'LoBrA'  Data Object (LDO) for Example data set}
\format{
A matrix representing 20 experiments. It contains values for 100 variables at 18 time points for each experiment.
Object of class \code{LDO}.
}
\usage{
ldo
}
\description{
'LoBrA'  example LDO created by the function 'createExampleData' and converted to an LDO by 'as.LOBdataset' function.
It consist of a single matrix for all experiments, time points, types (background, experiment), class and the intensity values of all components created.
The artificial data consist of 20 experiments and 100 components with 18 measurements (3 background, 15 sample). The 10 experiments are each associated to on of 2 groups (ONE and TWO). The components comprise 70 noise components and 30 components that randomly vary in their trajectories in one of three segments. Random noise is added to all intercepts, propagated and added to each time point for all samples and components separately.
}
\author{
Anne-Christin Hauschild \email{hauschild@uni-marburg.de}
}
\keyword{Breathomics}
\keyword{Confounders}
\keyword{Expression}
\keyword{Longitudinal-Data}
\keyword{Metabolomics}
