% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{LDOmodelselection-class}
\alias{LDOmodelselection-class}
\alias{LDOmodelselection}
\title{An S4 class to represent a model selection result based on an 'LDO'.}
\description{
An S4 class to represent a model selection result based on an 'LDO'.
}
\section{Slots}{

\describe{
\item{\code{ldo}}{LDO 'LDO' object the model selection is based on.}

\item{\code{potentialBreaks}}{numeric Vector of numeric values that were considered as potential break points in the model selection.}

\item{\code{splinetype}}{character Type of spline used.}

\item{\code{qualityMeasure}}{character Quality measures used during the model selection ('AIC', 'BIC' or 'LogLik')}

\item{\code{modelList}}{list List of all models tested.}

\item{\code{quality}}{list List of quality matrices, one matrix for each quality measure used. Each matrix contains the quality for each spline tested for each component.}

\item{\code{breaks}}{list For each tested spline, this list contains a vector of breaks.}
}}

