% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{LDOscreening-class}
\alias{LDOscreening-class}
\alias{LDOscreening}
\title{An S4 class to represent the screening of metabolites in an 'LDO'.}
\description{
An S4 class to represent the screening of metabolites in an 'LDO'.
}
\section{Slots}{

\describe{
\item{\code{ldo}}{LDO LDO object the screening is based on.}

\item{\code{experimentIntercept}}{list List of experiment intercepts.}

\item{\code{experimentResiduals}}{list List of experiment Residuals.}

\item{\code{interceptPvalues}}{matrix Matrix of experiment intercept p-values.}

\item{\code{residualPvalues}}{matrix Matrix of experiment Residual p-values.}

\item{\code{selectedPeaks}}{matrix Matrix of logical values. Each entry indicates whether a specific component is significant according to a specific test.}
}}

