\name{LoBrA-package}
\alias{LoBrA-package}
\alias{LoBrA}
\docType{package}
\title{
\packageTitle{LoBrA}
}
\description{
\packageDescription{LoBrA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LoBrA}
\packageIndices{LoBrA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LoBrA}

Maintainer: \packageMaintainer{LoBrA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
