% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.peaksplines.R
\name{plot.modelSelectionEvaluation}
\alias{plot.modelSelectionEvaluation}
\title{Plotting background screening results.}
\usage{
\method{plot}{modelSelectionEvaluation}(modelSelectionObject, qualityMeasure)
}
\arguments{
\item{modelSelectionObject}{Object of type LDOmodelselection that was created during the model evaluation. @seealso lobraModelSelection}

\item{qualityMeasure}{Quality measure to be visualized.}
}
\description{
Plotting the results of the background screening of backbround or confounding components.
}
\examples{
\dontrun{
} 
  

}
