% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudermanModelling.R
\name{plotGaudermanModel}
\alias{plotGaudermanModel}
\title{Plotting helper function to plot a single generalized gouderman Model}
\usage{
plotGaudermanModel(
  data,
  labels,
  ul,
  tempmodel,
  colores,
  maincol,
  breaks,
  main,
  ylab,
  xlab
)
}
\arguments{
\item{data}{data matrix used to fit the model}

\item{labels}{class labels for all samples}

\item{ul}{unique class labels}

\item{tempmodel}{model to be plotted}

\item{colores}{predefined colors for the single samples}

\item{maincol}{predefined colors for the fitted spline}

\item{breaks}{break points of the spline to be plotted}

\item{main}{main title of the plot}

\item{ylab}{y label of the plot}

\item{xlab}{x label of the plot}
}
\description{
Plotting helper function to plot a single generalized gouderman Model
}
