% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{plotScreening}
\alias{plotScreening}
\title{Plotting the screening results.}
\usage{
plotScreening(ldoscreen, plotAll = FALSE, correctionmethod = "levene",
  decs = 3)
}
\arguments{
\item{ldoscreen}{LDO screening result}

\item{plotAll}{Select all components to be plottet. Default plots only the selected peaks using the correctionmethod.}

\item{correctionmethod}{Version of correction method to be used to select the peaks. Valid values are "bf", "levene", and "bartlett".}

\item{decs}{decimal numbers of p-values to be plotted.}
}
\value{
LDO data object
}
\examples{
\dontrun{
} 

data(ldosExample)
correctionmethod= "levene"

filename<-paste("screeningresults.pdf",sep="") 
pdf(filename, width=12, height=6)
plotScreening(ldoscreen)
dev.off();

}
