% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plotting function for a longitudinal data matrix (Internal Function)}
\usage{
plotTimeSeries(
  myMatrix,
  main = "",
  labels = NA,
  ylab = "Expression",
  xlab = "Time Point",
  legend = "",
  col = 1:dim(myMatrix)[1]
)
}
\arguments{
\item{myMatrix}{longitudinal data matrix to be plotted}

\item{main}{Title of the plot}

\item{labels}{class labels of samples}

\item{ylab}{Label of y axis}

\item{xlab}{Label of x axis}

\item{legend}{of plot}

\item{col}{vector of colors for plot}
}
\value{
No return value
}
\description{
Plotting function for a longitudinal data matrix (Internal Function)
}
