% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.peaksplines.R
\name{powerSet}
\alias{powerSet}
\title{Creating the power set of a set.}
\usage{
powerSet(set)
}
\arguments{
\item{set}{Set of numbers of potential spline break points.}
}
\value{
Returns power set of the given set.
}
\description{
Creating the power set of a set.
}
