% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{createBGComponents}
\alias{createBGComponents}
\title{Create example background noise peaks}
\usage{
createBGComponents(components = 50, samples = 10, labels = NA,
  timepoints = 15, bg = 3, mean = 5, sd = 3, experimentSD = 2,
  randomnoise = 0.1, plotting = FALSE)
}
\arguments{
\item{samples}{number of experiments}

\item{timepoints}{number of sample measurements}

\item{bg}{number of background measurements}

\item{mean}{mean value of noise components}

\item{sd}{standard deviation value of noise for this component}

\item{randomnoise}{random variation changing at each time point}

\item{component}{number of background components to be created}

\item{label}{label of each experiment}

\item{experiementSD}{standard deviation value of each experiment for this component}
}
\value{
matrix of background components
}
\description{
Background noise signals originating from experimental settings or random events
}
