% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudermanModelling.R
\name{createGoudermanData}
\alias{createGoudermanData}
\title{Create the Gouderman Data and Model}
\usage{
createGoudermanData(selectedLDO, breaks, center, timeperiod = NA,
  range = NA)
}
\arguments{
\item{s}{A numeric value to}
}
\value{
square of the in put
}
\description{
The screening method extracts peaks that are likely to
}
\examples{
\dontrun{
} 
  data(modelSelectionObjectExample)
  
  optimal<-getOptimalSpline(lobraModelSelectionObject, qualityMeasure="AIC")
  breaks<- optimal@breaks$optimalID
  center<- 14
  timeperiod <- 4;
  gaudermanLDOexample4TP <- createGoudermanData(modelSelectionObject@ldo, breaks, center, timeperiod)
  save(gaudermanLDOexample4TP,file="data/mygaudermanLDOExample4.RData")
  
  range <- c(9,13)
  center<- 9;
  gaudermanLDOexample3TP <- createGoudermanData(modelSelectionObject@ldo, breaks=range, center, range=range);
  save(gaudermanLDOexample3TP,file="data/mygaudermanLDOExample3.RData")
  
}
