% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{createInformativeComponents}
\alias{createInformativeComponents}
\title{Create example informative peaks}
\usage{
createInformativeComponents(components = 10, samples = 10, labels = NA,
  timepoints = c(5, 5, 5), bg = 3, mean = 5, sd = 3, segment = 1,
  myl = labels[1], slopeSD = 2, randomnoise = 0.5, plotting = FALSE)
}
\arguments{
\item{samples}{number of experiments}

\item{timepoints}{number of sample measurements}

\item{bg}{number of background measurements}

\item{mean}{mean value of noise components}

\item{sd}{standard deviation value of noise for this component}

\item{randomnoise}{random variation changing at each time point}

\item{component}{number of background components to be created}

\item{label}{label of each experiment}

\item{experiementSD}{standard deviation value of each experiment for this component}
}
\value{
matrix of background components
}
\description{
Background noise signals originating from experimental settings or random events
}
