% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{GaudermanLDO-class}
\alias{GaudermanLDO-class}
\alias{GaudermanLDO}
\title{An S4 class to represent a gouderman LDO object, that was generated by the generalized gauderman algorithm.}
\description{
An S4 class to represent a gouderman LDO object, that was generated by the generalized gauderman algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{character}

\item{\code{dataFrames}}{list}

\item{\code{peaknames}}{character}

\item{\code{k}}{numeric}

\item{\code{times}}{matrix",newTimeVars    character}

\item{\code{ids}}{character}

\item{\code{labels}}{factor"}
}}

