% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{GaudermanModelEvaluation-class}
\alias{GaudermanModelEvaluation-class}
\alias{GaudermanModelEvaluation}
\title{An S4 class to represent the result of the linear mixed effect modeling on a gauderman LDO.}
\description{
An S4 class to represent the result of the linear mixed effect modeling on a gauderman LDO.
}
\section{Slots}{

\describe{
\item{\code{name}}{character}

\item{\code{gaudermanLDO}}{GaudermanLDO}

\item{\code{models}}{list}

\item{\code{pvalues}}{matrix}

\item{\code{correctedpvalues}}{matrix}

\item{\code{modelparameter}}{matrix"))#'}
}}

