% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.peaksplines.R
\name{getOptimalSpline}
\alias{getOptimalSpline}
\title{Extract the optimal spline model parameters from the ModelSelection Object.

 @description The method calculates which spline model and parameters worked best over all different components.}
\usage{
getOptimalSpline(modelSelectionObject, qualityMeasure = "AIC")
}
\arguments{
\item{qualityMeasure}{Quality measure to be used to select the optimal spline.}
}
\value{
modelSelectionObject that contains the optimal model.
}
\examples{
\dontrun{
} 
  data(modelSelectionObjectExample)
  
  optimal<-getOptimalSpline(modelSelectionObject, qualityMeasure="AIC")

}
