% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{LDO-class}
\alias{LDO-class}
\alias{LDO}
\title{An S4 class to represent a LoBra Data Object (LDO).}
\description{
An S4 class to represent a LoBra Data Object (LDO).
}
\section{Slots}{

\describe{
\item{\code{name}}{character}

\item{\code{dataMatrices}}{list}

\item{\code{backgroundMatrices}}{list}

\item{\code{peaknames}}{character}

\item{\code{times}}{numeric}

\item{\code{ids}}{character}

\item{\code{labels}}{factor}
}}

