% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{LDOmodelselection-class}
\alias{LDOmodelselection-class}
\alias{LDOmodelselection}
\title{An S4 class to represent a model selection result based on an LDO.}
\description{
An S4 class to represent a model selection result based on an LDO.
}
\section{Slots}{

\describe{
\item{\code{ldo}}{LDO}

\item{\code{potentialSlines}}{numeric}

\item{\code{splinetype}}{character}

\item{\code{qualityMeasure}}{character}

\item{\code{modelList}}{list}

\item{\code{quality}}{list}

\item{\code{breaks}}{list"))}
}}

