% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\docType{class}
\name{LDOscreening-class}
\alias{LDOscreening-class}
\alias{LDOscreening}
\title{An S4 class to represent the screening of metabolites in an LDO.}
\description{
An S4 class to represent the screening of metabolites in an LDO.
}
\section{Slots}{

\describe{
\item{\code{ldo}}{LDO}

\item{\code{experimentIntercept}}{list}

\item{\code{experimentResiduals}}{list}

\item{\code{interceptPvalues}}{matrix}

\item{\code{residualPvalues}}{matrix}

\item{\code{selectedPeaks}}{matrix"))}
}}

