% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.peaksplines.R
\name{lobraModelSelection}
\alias{lobraModelSelection}
\title{Evaluation of different spline variants.}
\usage{
lobraModelSelection(selectedLDO, potentialSplines = c(), nknots = c(0, 1,
  2), splinetype = "linear", qualityMeasure = c("AIC", "BIC", "logLik"))
}
\arguments{
\item{selectedLDO}{LDO containing all selected metabolites to be used for the model selection.}

\item{potentialSplines}{Vector of all possible knots to be used for the spline modeling.}

\item{nknots}{Vector of number of spline knots to be used.}

\item{splinetype}{spline type default is "linear".}

\item{qualityMeasure}{Vector of quality measures to be used. Possible options are "AIC", "BIC", and "logLik".}
}
\value{
square of the in put
}
\description{
The model selection method evaluates which spline models achieve the best quality among all tested metabolites.
}
\examples{
\dontrun{
} 
  data(ldoExample)
  potentialSplines <- c(6,8,10,12,14,16)
  
  nknots=c(0,1, 2, 3, 4)
  splinetype="linear"
  qualityMeasure=c("AIC", "BIC", "logLik")
  components <- ldos@selectedPeaks[,"bf"]
  components <- names(components)[components]
  selectedLDO <- selectComponents(ldo, components)
  lobraModelSelectionObject<- lobraModelSelection(selectedLDO, potentialSplines, nknots, splinetype, qualityMeasure)
  # save(modelSelectionObject,file="data/modelSelectionObjectExample.RData")
  
  
}
