% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaudermanModelling.R
\name{modelGoudermanLongitudinal}
\alias{modelGoudermanLongitudinal}
\title{Create the Gouderman Model for each of the metabolites}
\usage{
modelGoudermanLongitudinal(mygaudermanLDO, splinetype = "linear")
}
\arguments{
\item{mygaudermanLDO}{GaudermanLDO data object, created by the generalized gauderman algorithm (GGA).}
}
\value{
GaudermanModelEvaluation Results of the evaluation of the Fitted linear mixed effect models for the defined time periods.
}
\description{
Fitting the Gouderman Model with using Gouderman-Data Arangement.
}
\examples{
\dontrun{
} 
  data(mygaudermanLDOExample3)
  evaluationresult1<- modelGoudermanLongitudinal(gaudermanLDOexample3TP)
  
  data(mygaudermanLDOExample4)
  evaluationresult1<- modelGoudermanLongitudinal(gaudermanLDOexample4TP)
  
}
