% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initExampleData.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plotting Function for a longitudinal data matrix}
\usage{
plotTimeSeries(myMatrix, main = "", labels = NA, ylab = "Expression",
  xlab = "Time Point", legend = "", col = 1:dim(myMatrix)[1])
}
\arguments{
\item{myMatrix}{matrix to be plotted}

\item{labels}{class labels of samples}

\item{ylab}{Label of y achis}

\item{xlab}{Label of x achis}

\item{legend}{of plot}

\item{col}{vector of colors for plot}
}
\description{
Plotting Function for a longitudinal data matrix
}
