% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.peaksplines.R
\name{powerSet}
\alias{powerSet}
\title{Creating the power set of a set..}
\usage{
powerSet(set)
}
\arguments{
\item{set}{Set of numbers or strings.}
}
\value{
Returns powerset of the given set.
}
