% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoBrA.R
\name{selectComponents}
\alias{selectComponents}
\title{lobra functionality}
\usage{
selectComponents(ldo, components)
}
\arguments{
\item{ldo}{Longitudinal Data Object}

\item{components}{component names to select for the new ldo object. Only elements from this list that overlab with the peaknames in the given ldo, are utilized.}
}
\value{
new ldo object only containing the defined components.
}
\description{
Does this and that
}
